requirements.txt
pillow==12.2.0
PyYAML==6.0.3
